+++ src/pytest_executable/script_runner.py
@@ -15,14 +15,15 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-"""This module contains functions to execute a script in a sub process."""
+"""Provides the shell script creation and execution routines."""
 
 import logging
 import stat
 import subprocess
 from pathlib import Path
-from typing import Any, Dict, Iterator
+from typing import Dict
 
+import delta
 import jinja2
 
 LOG = logging.getLogger(__name__)
@@ -33,100 +34,115 @@ class ScriptExecutionError(Exception):
 
 
 class ScriptRunner:
-    """Base class for creating and running scripts.
+    """Class for creating and running a runner script.
 
-    Args:
-        name: Name of the script.
-        script: Content of the script.
-        exec_dir: Path to the execution directory.
+    It can create and execute a shell script from the path to a script with
+    placeholders and the settings.
+
+    Attributes:
+        path: Path to the script.
+        settings: Runner settings from the yaml file.
+        workdir: Path to the script working directory.
+        STDOUT_EXT: Suffix for the file with the script standard output (class
+                    attribute).
+        STDERR_EXT: Suffix for the file with the script standard error (class
+                    attribute).
+        SHELL: Shell used to execute the script (class attribute).
     """
 
-    # extensions for the script related files
-    SCRIPT_EXT = "sh"
     STDOUT_EXT = "stdout"
     STDERR_EXT = "stderr"
-
-    # shell used for executing the shell script
     SHELL = "/usr/bin/env bash"
 
-    def __init__(self, name: str, script: str, exec_dir: Path):
-        self._name = name
-        self._script = script
-        self._exec_dir = exec_dir
+    def __init__(self, path: Path, settings: Dict[str, str], workdir: Path):
+        """Docstring just to prevent the arguments to appear in the autodoc.
 
-    def _get_filenames(self) -> Iterator[str]:
-        """Return the script, stdout and stderr filenames."""
-        return map(
-            lambda s: "{}.{}".format(self._name, s),
-            (self.SCRIPT_EXT, self.STDOUT_EXT, self.STDERR_EXT),
-        )
+        Args:
+            path: Path to the script.
+            settings: Runner settings from the yaml file.
+            workdir: Path to the script working directory.
+        """
+        self.path = path
+        self.workdir = workdir
+        self.settings = settings
+        self._content = self._substitute()
+
+    def _substitute(self) -> str:
+        """Return the script contents with replaced placeholders.
+
+        Returns:
+            The final script contents.
+
+        Raises:
+            TypeError: if the script cannot be processed.
+            ValueError: if a placeholder is undefined.
+        """
+        try:
+            template = jinja2.Template(
+                self.path.read_text(), undefined=jinja2.StrictUndefined
+            )
+        except UnicodeDecodeError:
+            raise TypeError(f"cannot read the script {self.path}") from None
+        try:
+            return template.render(**self.settings)
+        except jinja2.exceptions.UndefinedError as error:
+            raise ValueError(f"in {self.path}: {error}") from None
 
     def run(self) -> int:
         """Execute the script.
 
-        The script file is created in the execution directory by dumping the
-        contents of the script string.
-        The script stdout and stderr are each redirected to files named after
-        the script file prefix and suffixed with stdout and stderr.
+        The script is created and executed in the working directory. The stdout
+        and stderr of the script are each redirected to files named after the
+        script and suffixed with :py:data:`STDOUT_EXT` and :py:data:`STDERR_EXT`.
 
         Returns:
             The return code of the executed subprocess.
 
         Raises:
-            ScriptExecutionError: If the process fails.
+            ScriptExecutionError: If the execution fails.
         """
-        script_filename, stdout_filename, stderr_filename = self._get_filenames()
-
-        script_path = self._exec_dir / script_filename
+        filename = self.path.name
+        script_path = self.workdir / filename
 
         # write the script
         with script_path.open("w") as script_file:
             LOG.debug("writing the shell script %s", script_path)
-            script_file.write(self._script)
+            script_file.write(self._content)
 
         # make it executable for the user and the group
         permission = stat.S_IMODE(script_path.stat().st_mode)
         script_path.chmod(permission | stat.S_IXUSR | stat.S_IXGRP)
 
         # redirect the stdout and stderr to files
-        stdout = open(self._exec_dir / stdout_filename, "w")
-        stderr = open(self._exec_dir / stderr_filename, "w")
+        stdout = open(self.workdir / f"{filename}.{self.STDOUT_EXT}", "w")
+        stderr = open(self.workdir / f"{filename}.{self.STDERR_EXT}", "w")
 
         LOG.debug("executing the shell script %s", script_path)
-        cmd = self.SHELL.split() + [script_filename]
+        cmd = self.SHELL.split() + [filename]
+
+        timeout = self.settings.get("timeout")
+
+        if timeout is not None:
+            # convert to seconds
+            timeout = delta.parse(timeout).seconds
 
         try:
             process = subprocess.run(
-                cmd, cwd=self._exec_dir, stdout=stdout, stderr=stderr, check=True
+                cmd,
+                cwd=self.workdir,
+                stdout=stdout,
+                stderr=stderr,
+                check=True,
+                timeout=timeout,  # type: ignore
             )
         except subprocess.CalledProcessError:
             # inform about the log files
             raise ScriptExecutionError(
                 "execution failure, see the stdout and stderr files in "
-                f"{self._exec_dir}"
+                f"{self.workdir}"
             )
         finally:
             stdout.close()
             stderr.close()
 
         return process.returncode
-
-
-def get_final_script(script_path: Path, variables: Dict[str, Any]) -> str:
-    """Return the script with replaced placeholders.
-
-    Args:
-        script_path: Path to the script.
-        variables: Placeholders to be replaced.
-
-    Returns:
-        The final script.
-
-    Raises:
-        TypeError if the script cannot be processed.
-    """
-    try:
-        template = jinja2.Template(script_path.read_text())
-    except UnicodeDecodeError:
-        raise TypeError(f"can't read the script {script_path}") from None
-    return template.render(**variables)