+++ tests/data/runners/error.sh
@@ -1,0 +1,1 @@
+ls non-existing-file